# Optimization batch 9: avoid detecting irrelevant renames