This series depends on ????.

This series determines paths which meet special cases where detection
of renames is irrelevant, where the irrelevance is due to the fact
that the merge machinery will arrive at the same result regardless of
whether a rename is detected for any of those paths.  It represents
"Optimization #2" from my Git Merge 2020 talk[1], though this series
has some improvements on the optimization relative to what I had at
that time.  For the testcases mentioned in commit ??????????
("merge-ort: begin performance work; instrument with trace2_region_*
calls", 2020-10-28), the changes in just this series improves the
performance as follows:

                         Before Series           After Series
    no-renames:      12.544 s ±  0.069 s      5.530 s ±  0.038 s
    mega-renames:   123.506 s ±  0.372 s     13.485 s ±  0.139 s
    just-one-mega:    3.757 s ±  0.011 s    491.5  ms ±  3.7  ms

However, interestingly, if we had ignored the last two optimization
batches doing basename-guided rename detection (en/?????? and
en/?????), then this optimization series would have improved the
performance as follows:

                   Before Basename Series   After Just This Series
    no-renames:      13.815 s ±  0.062 s      5.728 s ±  0.104 s
    mega-renames:  1799.937 s ±  0.493 s     18.213 s ±  0.139 s
    just-one-mega    51.289 s ±  0.019 s    891.9  ms ±  7.0  ms

In other words, this optimization nearly renders unnecessary the
basename-guided rename detection from the last two series *IF* one
side of history doesn't modify very many files that the other side
renames (something that is particularly true for rebase testcases
since each patch typically only modifies a handful of files).  In
effect, the two optimizations compete and it is common for both kinds
of optimizations to be possible, but we will only get the speedups
from the better of the two (which the last patch in this series makes
pretty clear).  But, it is helpful to know that in alternative
testcases where this irrelevant-renames-optimization doesn't apply to
as many paths, that the basename-guided rename detection optimizations
can still often provide significant speedups for those paths.

As a reminder, before any merge-ort/diffcore-rename performance work,
the performance results we started with (as noted in the same commit
message) were:

    no-renames-am:      6.940 s ±  0.485 s
    no-renames:        18.912 s ±  0.174 s
    mega-renames:    5964.031 s ± 10.459 s
    just-one-mega:    149.583 s ±  0.751 s

The third-to-last patch of this series also discusses some minor
variants of the irrelevant-renames-optimization that were tried (and
resulted in less effective performance gains than reported here), as
well as some of the preparatory work over the past few years that this
series relies on in order to enable this optimization.

The eagle-eyed will note that the rows labelled "everything else" from
the commit message of 36d1f87d05 represent the maximum possible
speed-up from accelerating rename detection alone; basically, those
rows represent how fast the code could be if we had somehow infinitely
parallelized the inexact rename detection.  However, if you compare
those numbers to what we have above, you'll note that the infinitely
parallelized inexact rename detection would have been slightly
*slower* than the results we have now achieved.  (The reason this is
possible, despite the fact that we still spend time in rename
detection after our optimizations, is because we implemented two
optimizations outside of diffcore_rename() along the way.)  However,
this good news does also come with a downside -- it means that our
remaining optimization potential is somewhat limited, and subsequent
optimization series will have to fight for much smaller gains.

[1] https://github.com/newren/presentations/blob/pdfs/merge-performance/merge-performance-slides.pdf